Draft: Expand package toml

See merge request redox-os/pkgutils!70

+++ pkg-lib/src/callback/mod.rs
@@ -7,7 +7,7 @@ mod plain;
 
 pub trait Callback {
     fn start_download(&mut self, length: u64, file: &str);
-    fn increment_downloaded(&mut self, downloaded: usize);
+    fn increment_downloaded(&mut self, downloaded: u64);
     fn end_download(&mut self);
 
     fn conflict(&mut self) {}
